requirements.txt
spacy==3.8.4
wikipedia==1.4.0
tavily-python==0.5.1
openai==1.65.4
python-dotenv==1.0.1
langchain-community==0.3.19
langchain-anthropic==0.3.9
requests==2.32.3
pydantic==2.10.6
typing_extensions==4.12.2
langgraph==0.3.5
langchain==0.3.20
langchain-core==0.3.43
langchain-openai==0.3.8
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
en-core-web-trf==3.8.0
torch>=2.6.0
spacy-curated-transformers>=0.3.0
curated-transformers>=0.1.1
curated-tokenizers>=0.0.9